requirements.txt
Django==1.10.3
argparse==1.2.1
arrow==0.9.0
distribute==0.6.24
djangorestframework==3.5.3
python-dateutil==2.6.0
requests==2.12.1
six==1.10.0
wsgiref==0.1.2
